+++ example/lib/main.dart
@@ -1,4 +1,3 @@
-import 'package:example/material3.dart';
 import 'package:flutter/material.dart';
 
 import 'example.dart';
@@ -11,18 +10,16 @@ class MyApp extends StatelessWidget {
   const MyApp();
 
   build(context) {
-    return Material3App();
-
-    // return MaterialApp(
-    //   title: "Niku",
-    //   debugShowCheckedModeBanner: false,
-    //   theme: ThemeData(
-    //     primaryColor: Colors.blue,
-    //     primarySwatch: Colors.blue,
-    //     useMaterial3: true,
-    //   ),
-    //   darkTheme: ThemeData.dark(),
-    //   home: const Tempestissimo(),
-    // );
+    return MaterialApp(
+      title: "Niku",
+      debugShowCheckedModeBanner: false,
+      theme: ThemeData(
+        primaryColor: Colors.blue,
+        primarySwatch: Colors.blue,
+        useMaterial3: true,
+      ),
+      darkTheme: ThemeData.dark(),
+      home: AegleSeeker(),
+    );
   }
 }